Split the makerdao module into DSR and Vaults modules
