+++ rotkehlchen/chain/ethereum/makerdao/__init__.py
@@ -1,0 +1,2 @@
+from rotkehlchen.chain.ethereum.makerdao.dsr import MakerDAODSR  # noqa: F401
+from rotkehlchen.chain.ethereum.makerdao.vaults import MakerDAOVaults  # noqa: F401
